+++ frontend/src/business/services/SocketManager/__mocks__/HumanSocketManager.ts
@@ -1,10 +1,10 @@
-export default {
-  getInstance: vi.fn().mockReturnValue({
+export class HumanSocketManager {
+  static getInstance = vi.fn().mockReturnValue({
     connect: vi.fn(),
     disconnect: vi.fn(),
     on: vi.fn(),
     emit: vi.fn(),
     socket: {},
     connected: false,
-  }),
-};
+  });
+}
