+++ frontend/src/business/services/SocketManager/index.ts
@@ -1,3 +1,3 @@
-export { default } from './SocketManager';
-export { default as AISocketManager } from './AISocketManager';
-export { default as HumanSocketManager } from './HumanSocketManager';
+export * from './SocketManager';
+export * from './AISocketManager';
+export * from './HumanSocketManager';
